frist change on v1.1, remove the backward reliance from modbus rtu to modbus controller (timer)

+++ modbus_controller/modbus_rtu/Include/port.h
@@ -49,7 +49,7 @@ typedef unsigned char   UCHAR;
 typedef char            CHAR;
 typedef uint16_t        USHORT;
 typedef int16_t         SHORT;
-typedef uint32_t        ULONG;
+typedef unsigned long   ULONG;
 typedef int32_t         LONG;
 typedef uint8_t         UINT8;
